requirements.txt
# ============================================================
# requirements.txt
# Global AI & Economic Data Aggregator — MSBA 305
# ============================================================
#
# Install all dependencies:
#   pip install -r requirements.txt
#
# Python version: 3.11+ recommended (tested on 3.11)
# ============================================================


# ─────────────────────────────────────────────────────────────
# DATABASE
# ─────────────────────────────────────────────────────────────
sqlalchemy>=2.0.0          # ORM and database connection layer
psycopg2-binary>=2.9.0     # PostgreSQL adapter for SQLAlchemy
                           # binary version — no system deps needed


# ─────────────────────────────────────────────────────────────
# CLOUD STORAGE (Backblaze B2)
# ─────────────────────────────────────────────────────────────
boto3>=1.34.0              # S3-compatible client for Backblaze B2
botocore>=1.34.0           # boto3 dependency (pinned for compatibility)


# ─────────────────────────────────────────────────────────────
# DATA PROCESSING
# ─────────────────────────────────────────────────────────────
pandas>=2.0.0              # DataFrames — core data manipulation
numpy>=1.26.0              # Numeric operations (pandas dependency)
openpyxl>=3.1.0            # Read/write .xlsx files (Oxford source)
pyreadstat>=1.2.0          # Read Stata .dta files (PWT source)


# ─────────────────────────────────────────────────────────────
# HTTP / API CLIENTS
# ─────────────────────────────────────────────────────────────
requests>=2.31.0           # HTTP client for World Bank, IMF,
                           # OpenAlex, OECD API calls
wbgapi>=1.0.12             # World Bank API wrapper
                           # used in seed scripts only


# ─────────────────────────────────────────────────────────────
# COUNTRY CODE UTILITIES
# ─────────────────────────────────────────────────────────────
pycountry>=22.3.5          # ISO country code lookups
                           # used in seed scripts and Oxford matching


# ─────────────────────────────────────────────────────────────
# ORCHESTRATION
# ─────────────────────────────────────────────────────────────
prefect>=2.19.0            # Pipeline orchestration and scheduling
                           # flows, tasks, cron scheduling


# ─────────────────────────────────────────────────────────────
# ENVIRONMENT / CONFIG
# ─────────────────────────────────────────────────────────────
python-dotenv>=1.0.0       # Load .env files into os.environ


# ─────────────────────────────────────────────────────────────
# STREAMLIT APP
# ─────────────────────────────────────────────────────────────
streamlit>=1.35.0          # Researcher-facing web application
plotly>=5.22.0             # Interactive charts in Streamlit
